adjusted cooldowns to be accurate added ult skill boosts for fun some typo fixes

+++ src/com/ebicep/warlords/classes/abilties/Breath.java
@@ -16,7 +16,7 @@ import java.util.List;
 
 public class Breath extends AbstractAbility {
 
-    public Breath(String name, float minDamageHeal, float maxDamageHeal, int cooldown, int energyCost, int critChance, int critMultiplier) {
+    public Breath(String name, float minDamageHeal, float maxDamageHeal, float cooldown, int energyCost, int critChance, int critMultiplier) {
         super(name, minDamageHeal, maxDamageHeal, cooldown, energyCost, critChance, critMultiplier);
     }
