+++ src/com/ebicep/warlords/classes/abilties/GroundSlam.java
@@ -22,7 +22,7 @@ public class GroundSlam extends AbstractAbility {
     private Player owner;
     private List<Player> playersHit = new ArrayList<>();
 
-    public GroundSlam(String name, float minDamageHeal, float maxDamageHeal, int cooldown, int energyCost, int critChance, int critMultiplier, Player owner) {
+    public GroundSlam(String name, float minDamageHeal, float maxDamageHeal, float cooldown, int energyCost, int critChance, int critMultiplier, Player owner) {
         super(name, minDamageHeal, maxDamageHeal, cooldown, energyCost, critChance, critMultiplier);
         this.owner = owner;
     }
